trans/models/trainModel/mainLib.py
import os
import cv2
import sys
import json
import glob
import time
import math
import yaml
import torch
import pickle
import numbers
import imutils
import argparse
import matplotlib
import numpy as np
import pandas as pd
import torch.nn as nn
from tqdm import tqdm
from pathlib import Path
import torch.optim as optim
from pandas import ExcelWriter
import matplotlib.pyplot as plt
from operator import itemgetter
import torch.nn.functional as F
from collections import namedtuple
from yaml.loader import SafeLoader
from collections import namedtuple
# from got10k.trackers import Tracker
from torch.utils.data import Dataset
from easydict import EasyDict as edict
from torch.utils.data import DataLoader
# from shapely.geometry import Polygon, box
from torch.optim.lr_scheduler import ExponentialLR